Self PR: skip Slack notify without a webhook; clarify CI and test mock guidance
Skip community Slack posts when the webhook secret is empty (forks / unset),
document when CI jobs run, and rename the events PocketBase test helper so
contributors know when to use it vs a per-file mock.
