+++ .github/README.md
@@ -2,6 +2,15 @@
 
 Repo admins configure secrets and workflows here. Contributors do not need this for local development.
 
+## When jobs run
+
+See [`workflows/ci.yaml`](workflows/ci.yaml):
+
+- **Vitest and Playwright** run on `pull_request` only.
+- **Docker build / deploy** runs on push to `main` or `dev` only — those pushes do not re-run the test jobs (tests are assumed green from the merged PR).
+
+Direct pushes that bypass a PR therefore also bypass Vitest/Playwright.
+
 ## PR notifications (`#community-projects`)
 
 When a pull request targets `dev`, is not a draft, and both CI test jobs (Vitest and Playwright) pass, the `notify-community-projects` job in [`workflows/ci.yaml`](workflows/ci.yaml) posts once to `#community-projects`.
@@ -22,7 +31,7 @@ When a pull request targets `dev`, is not a draft, and both CI test jobs (Vitest
 
 ### Limitations
 
-- **Fork PRs:** GitHub does not expose repository secrets to `pull_request` workflows from forks. Tests still run; Slack notification is skipped. Same-repo branches work normally.
+- **Fork PRs / missing secret:** GitHub does not expose repository secrets to `pull_request` workflows from forks, so the webhook env is empty. The notify job succeeds but skips the Slack post (same when `SLACK_COMMUNITY_PROJECTS_WEBHOOK` is unset). Same-repo branches with the secret configured still fail the job if `curl` errors — that is a real delivery failure.
 - **Cache expiry:** inactive PR caches may expire after ~7 days; a very stale PR could notify again on the next green CI run.
 
 ## Deploy notifications
