+++ CONTRIBUTING.md
@@ -18,7 +18,7 @@ The [#community-projects Website Project Board](https://github.com/orgs/indyhack
 
 - Smaller PRs are easier to review, but use whatever scope fits your change.
 - Run tests locally before opening a PR — commands in [README.md](README.md#run-tests-with-vitest).
-- Ensure CI passes on your PR — see [`.github/workflows/ci.yaml`](.github/workflows/ci.yaml) for what runs when.
+- Ensure CI passes on your PR — see [`.github/workflows/ci.yaml`](.github/workflows/ci.yaml) for what runs when. Vitest and Playwright run on pull requests only; pushes to `main`/`dev` build and deploy without re-running those tests (details: [`.github/README.md`](.github/README.md#when-jobs-run)).
 - PRs targeting `dev` that pass CI notify `#community-projects` once. Maintainers: see [`.github/README.md`](.github/README.md) for webhook setup.
 
 ## Testing
@@ -32,11 +32,11 @@ Vitest and Playwright currently use MSW — no Docker or PocketBase container is
 ### Vitest — default for most tests
 
 - **Where:** co-locate `*.test.js` beside the file under test under `src/` (e.g. `LoginPage.vue` + `LoginPage.test.js`).
-- **How:** runs in jsdom with MSW ([`vitest.setup.js`](vitest.setup.js)) or `fakePocketBase()` for isolated components.
+- **How:** runs in jsdom with MSW ([`vitest.setup.js`](vitest.setup.js)) or a PocketBase stand-in (`createFakeEventsPocketBase()` / a per-file mock) for isolated components.
 - **Use for:** utilities, composables, component render and interaction, form validation, API response handling, Pinia stores.
 - **Runs in CI** on every pull request (`vitest` job).
 
-Pure functions, single components, and anything you can mock with MSW or `fakePocketBase()` belong here. These tests should give sub-second feedback during development.
+Pure functions, single components, and anything you can mock with MSW or a PocketBase stand-in belong here. These tests should give sub-second feedback during development.
 
 ### Playwright — smoke and critical paths
 
@@ -47,7 +47,7 @@ Pure functions, single components, and anything you can mock with MSW or `fakePo
 - **CI:** `playwright` job on pull requests — Chromium only, same dev server + MSW setup as local.
 - **Debug:** `npm run test:e2e:ui` (interactive) or `npm run test:e2e:headed` (visible browser).
 
-The current smoke suite ([`e2e/smoke.spec.js`](e2e/smoke.spec.js)) covers five public routes: home, login, signup, jobs, and calendar. Keep e2e small and curated—a flaky or duplicated e2e test is worse than none.
+The smoke suite ([`e2e/smoke.spec.js`](e2e/smoke.spec.js)) covers a curated set of public routes — that file is the source of truth. At the time of writing: home, login, signup, jobs, calendar, and slack. Keep e2e small; a flaky or duplicated e2e test is worse than none.
 
 - **Runs in CI** on every pull request (`playwright` job, Chromium only).
 
@@ -56,7 +56,7 @@ The current smoke suite ([`e2e/smoke.spec.js`](e2e/smoke.spec.js)) covers five p
 | Question | If yes → |
 |----------|----------|
 | Pure function or single component? | Vitest |
-| Needs mocked PocketBase or API responses? | Vitest + MSW or `fakePocketBase()` |
+| Needs mocked PocketBase or API responses? | Vitest + MSW or a PocketBase stand-in |
 | Spans multiple routes or needs a real browser? | Playwright |
 | Would duplicate an existing Vitest test? | Don't add Playwright |
 
@@ -64,7 +64,19 @@ The current smoke suite ([`e2e/smoke.spec.js`](e2e/smoke.spec.js)) covers five p
 
 **Date-sensitive tests:** calendar helpers and views assume “today” is mid-June 2026 in Vitest. Use `vi.useFakeTimers()` and `vi.setSystemTime(new Date('2026-06-14T12:00:00-04:00'))` when testing upcoming-event logic (see `useEvents.test.js`). Playwright calendar smoke asserts page shell and static topic filters from MSW—not specific upcoming event titles.
 
-**PocketBase in component tests:** use `fakePocketBase()` from [`src/mocks/fakePocketBase.js`](src/mocks/fakePocketBase.js) for calendar-related components, or mock `$pocketbase` / `inject('pocketbase')` as other job tests do.
+**PocketBase in component tests:** how the component reads PocketBase decides the mount wiring; how much shared seed data you need decides the mock factory.
+
+| Component style | Wire the fake via |
+|-----------------|-------------------|
+| Composition API (`inject('pocketbase')`) | `global.provide: { pocketbase: … }` |
+| Options API (`this.pocketbase`) | `global.config.globalProperties.pocketbase` |
+
+| Need | Prefer |
+|------|--------|
+| Calendar/events UI backed by [`eventMocks.js`](src/mocks/eventMocks.js) | [`createFakeEventsPocketBase()`](src/mocks/createFakeEventsPocketBase.js) — events/topics/subscriptions only |
+| Specific return values for assertions (jobs, etc.) | Small per-file `createMockPocketBase()` (see `JobListing.test.js` / `JobsList.test.js`) |
+
+Do not grow `createFakeEventsPocketBase` into a general PocketBase fake. New features get their own helper or an inline mock.
 
 **Playwright selectors:** prefer `getByRole` and `getByText` (user-visible, resilient to refactors). Use `#id` locators only where forms already expose stable ids. Always use relative URLs (`page.goto('/jobs')`)—never hardcode ports; Playwright's `baseURL` in [`playwright.config.js`](playwright.config.js) handles that.
